价格调整

+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/converter/OrderVo2OutBoundConverter.java
@@ -105,7 +105,7 @@ public class OrderVo2OutBoundConverter implements Converter<List<OrderInfo>, Lis
                     try {
                         //计算不含税单价
                         Long aLong = map.get(vo.getSkuCode());
-                        BigDecimal noTaxPrice = Calculate.computeNoTaxPrice(vo.getPrice(), aLong);
+                        BigDecimal noTaxPrice = Calculate.computeNoTaxPrice(vo.getPrice(), BigDecimal.valueOf(aLong));
 
                         //计算不含税总价 (现在是主单位数量 * 单价）
 //                long noTaxTotalPrice = noTaxPrice * o.getNum();
