+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/converter/allocation/AllocationOrderToInboundConverter.java
@@ -119,7 +119,7 @@ public class AllocationOrderToInboundConverter implements Converter<AllocationDT
             preInboundNum += record.getQuantity().intValue();
             preInboundMainNum += record.getQuantity().intValue();
             preTaxAmount = record.getTaxAmount().add(preTaxAmount);
-            preNoTaxAmount = Calculate.computeNoTaxPrice(record.getTaxAmount(), record.getTax().longValue()).add(preNoTaxAmount);
+            preNoTaxAmount = Calculate.computeNoTaxPrice(record.getTaxAmount(), BigDecimal.valueOf(record.getTax().longValue()).add(preNoTaxAmount));
             list.add(reqVo1);
         }
         List<AllocationProductBatch> list1 = order.getList();
