+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/converter/supervisorywarehouseorder/WarehouseOrderToOutboundConverter.java
@@ -102,7 +102,7 @@ public class WarehouseOrderToOutboundConverter implements Converter<SupervisoryW
             preInboundNum += record.getNum().intValue();
             preInboundMainNum += record.getSingleCount().intValue();
             preTaxAmount = record.getProductTotalAmount().add(preTaxAmount);
-            preNoTaxAmount = Calculate.computeNoTaxPrice(record.getProductTotalAmount(), record.getTaxRate().longValue()).add(preNoTaxAmount);
+            preNoTaxAmount = Calculate.computeNoTaxPrice(record.getProductTotalAmount(), BigDecimal.valueOf(record.getTaxRate().longValue()).add(preNoTaxAmount));
             list.add(reqVo1);
             outboundBatchReqVos.add(batch);
         }
