+++ src/main/java/com/aiqin/bms/scmp/api/purchase/domain/PurchaseOrderProduct.java
@@ -88,7 +88,7 @@ public class PurchaseOrderProduct {
 
     @ApiModelProperty(value="税率")
     @JsonProperty("tax_rate")
-    private Integer taxRate;
+    private BigDecimal taxRate;
 
     @ApiModelProperty(value="含税单价")
     @JsonProperty("product_amount")
